[1.3.0] 版本更新 1. 更新 EasyConfiguration 到 3.0.0 (破坏性更新)。 2. 添加 ConfiguredTitle 用于快捷向玩家发送title消息。 3. 为 ConfiguredItem 添加name和lore的原生params变量支持。 4. 添加 ProtocolLibHelper 用于支持多版本情况下的部分功能。(如需使用请安装 ProtocolLib 插件)

+++ craftbukkit/src/main/java/cc/carm/lib/configuration/craft/data/TitleConfig.java
@@ -34,9 +34,9 @@ public class TitleConfig {
     }
 
     public void send(@NotNull Player player,
-                     @Range(from = 0L, to = Long.MAX_VALUE) long fadeIn,
-                     @Range(from = 0L, to = Long.MAX_VALUE) long stay,
-                     @Range(from = 0L, to = Long.MAX_VALUE) long fadeOut,
+                     @Range(from = 0L, to = Long.MAX_VALUE) int fadeIn,
+                     @Range(from = 0L, to = Long.MAX_VALUE) int stay,
+                     @Range(from = 0L, to = Long.MAX_VALUE) int fadeOut,
                      @NotNull Map<String, Object> placeholders,
                      @Nullable TitleSendConsumer sendConsumer) {
         if (this.line1 == null && this.line2 == null) return;
